Routes and data definitions run.  Logins cross.

+++ api/models/region.js
@@ -12,7 +12,7 @@ module.exports = function(sequelize, DataTypes) {
 			allowNull: true
 		},
 		quota: { 
-			type: DataTypes.TINYINT,
+			type: DataTypes.INTEGER(4),
 			allowNull: true
 		},
 		archdiocese: { 
@@ -24,7 +24,7 @@ module.exports = function(sequelize, DataTypes) {
 			allowNull: true
 		},
 		sweeps: { 
-			type: DataTypes.SMALLINT,
+			type: DataTypes.INTEGER(6),
 			allowNull: true
 		}
 	});
